Current iteration of the project as of 2/14/2025.
Current iteration of the project as of 2/14/2025.
Currently includes archetypes, abilities, mystical powers, rituals, attributes, races, traits, banes and boons. Also has the weapon qualities.

There are many things that need work still or are missing out right. There are place holders for the adventures to be incorporated into this database.

There is formatting to be done as well as other forms of editing and back linking between the files.